change avatar button connected!
Info popup needs to be worked on.

And the change avatar doesn't actually change anything rn. It's j a radio button. Was thinking about making it a "input type=file" maybe later. Idk how we want to save that. It's probably an extra line in js for local storage if we want to do that.

The info popoup doesnt work. and idk how to make it work tbh. i have some commented out js code that looks like it might work, but it like breaks the switching on the left.